+++ komorebic/src/main.rs
@@ -801,6 +801,9 @@ struct Stop {
     /// Stop masir if it is running as a background process
     #[clap(long)]
     masir: bool,
+    /// Do not restore windows after stopping komorebi
+    #[clap(long, hide = true)]
+    ignore_restore: bool,
 }
 
 #[derive(Parser)]
@@ -2300,7 +2303,11 @@ if (Get-Command Get-CimInstance -ErrorAction SilentlyContinue) {
                 }
             }
 
-            send_message(&SocketMessage::Stop)?;
+            if arg.ignore_restore {
+                send_message(&SocketMessage::StopIgnoreRestore)?;
+            } else {
+                send_message(&SocketMessage::Stop)?;
+            }
             let mut system = sysinfo::System::new_all();
             system.refresh_processes(ProcessesToUpdate::All);
